+++ src/main/java/site/metacoding/firstapp/service/LoveService.java
@@ -1,13 +1,10 @@
 package site.metacoding.firstapp.service;
 
-import java.util.List;
-
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
 import site.metacoding.firstapp.domain.love.Love;
 import site.metacoding.firstapp.domain.love.LoveDao;
-import site.metacoding.firstapp.web.dto.response.love.PostRespDto;
 
 @RequiredArgsConstructor
 @Service
@@ -26,11 +23,6 @@ public class LoveService {
 
 	public void 좋아요취소(Integer loveId) {
 		loveDao.delete(loveId);
-	}
-
-	public List<PostRespDto> 좋아요한게시글목록보기(Integer userId) {
-		List<PostRespDto> postRespDto = loveDao.findPostList(userId);
-		return postRespDto;
 	}
 
 }
